Fixed summary file output bug
Null pointer exceptions would happen when trying to close files due to a misguided attempt to create parent directories etc. This has been fixed and an error check has been added

+++ src/epics/camsim/core/Statistics.java
@@ -80,10 +80,11 @@ public class Statistics {
                 
                 if (summary != null && ! summary.equals("")) {
                 	File sumFile = new File(summary);
-                	File f1 = new File(sumFile.getParent());
-                	File f2 = new File(f1.getParent());
-                	boolean creat2 = f2.mkdir();
-                	boolean creat = f1.mkdir();
+                	File f1 = new File(sumFile.getAbsoluteFile().getParent());
+                	if (! f1.exists()) {
+                		throw new IOException("Error: Parent directory of summary file does not exist: "+
+                				f1.getAbsolutePath());
+                	}
                 	boolean existed = sumFile.exists();
                 	FileWriter sumFileWriter = new FileWriter(sumFile, true); // Append
                     sumOut = new PrintWriter(sumFileWriter);
